first working 64 bit version
--HG--
branch : 64bit

+++ src/memory/CopyingHeap.cpp
@@ -13,21 +13,21 @@ CopyingHeap::CopyingHeap(int objectSpaceSize) : Heap(objectSpaceSize) {
 	oldBuffer = malloc(bufSize);
 	memset(currentBuffer, 0x0, bufSize);
 	memset(oldBuffer, 0x0, bufSize);
-	currentBufferEnd = (void*)((int32_t)currentBuffer + bufSize);
-	collectionLimit = (void*)((int32_t)currentBuffer + ((int32_t)(bufSize *
+	currentBufferEnd = (void*)((size_t)currentBuffer + bufSize);
+	collectionLimit = (void*)((size_t)currentBuffer + ((size_t)(bufSize *
 					0.9)));
 	nextFreePosition = currentBuffer;
 }
 
 
 void CopyingHeap::switchBuffers() {
-	int32_t bufSize = (int32_t)currentBufferEnd - (int32_t)currentBuffer;
+	size_t bufSize = (size_t)currentBufferEnd - (size_t)currentBuffer;
 	void* tmp = oldBuffer;
 	oldBuffer = currentBuffer;
 	currentBuffer = tmp;
-	currentBufferEnd = (void*)((int32_t)currentBuffer + bufSize);
+	currentBufferEnd = (void*)((size_t)currentBuffer + bufSize);
 	nextFreePosition = currentBuffer;
-	collectionLimit = (void*)((int32_t)currentBuffer + (int32_t)(0.9 *
+	collectionLimit = (void*)((size_t)currentBuffer + (size_t)(0.9 *
 				bufSize));
 }
 
@@ -35,7 +35,7 @@ void CopyingHeap::switchBuffers() {
 AbstractVMObject* CopyingHeap::AllocateObject(size_t size) {
 	size_t paddedSize = size + PAD_BYTES(size);
 	AbstractVMObject* newObject = (AbstractVMObject*) nextFreePosition;
-	nextFreePosition = (void*)((int32_t)nextFreePosition + paddedSize);
+	nextFreePosition = (void*)((size_t)nextFreePosition + paddedSize);
 	if (nextFreePosition > currentBufferEnd) {
 		cout << "Failed to allocate " << size << " Bytes." << endl;
 		_UNIVERSE->Quit(-1);
