+++ src/vmobjects/VMString.h
@@ -44,7 +44,7 @@ public:
     virtual pVMString Clone() const;
 #endif
 	virtual pVMClass GetClass() const;
-	virtual int32_t GetObjectSize() const;
+	virtual size_t GetObjectSize() const;
 protected:
     //this could be replaced by the CHARS macro in VMString.cpp
     //in order to decrease the object size
